pyproject.toml
[project]
name = "raspilunii"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ctranslate2>=4.6.0",
    "faster-whisper>=1.2.0",
    "gpiozero>=2.0.1",
    "just-playback>=0.1.8",
    "numpy>=2.3.4",
    "ollama>=0.6.0",
    "pillow>=12.0.0",
    "pvrecorder>=1.2.7",
    "pyaudio>=0.2.14",
    "pynput>=1.8.1",
    "pyqt6>=6.10.0",
    "requests>=2.32.5",
    "spidev>=3.8",
    "types-requests>=2.32.4.20250913",
]
